fix: disabled camera collision testing to prevent camera jerking around
Disabling collision testing on the camera is a very hacky solution to a more complex problem, which is that pieces of scene geometry cause the camera to jerk around as the boom arm gets longer and shorter.  But for now, it's fine - my son won't notice that it's now possible to put the camera inside rocks, and peer inside buildings you shouldn't be able to, etc.