feat: Naive implementation of Common UI
Note that button (not axis) inputs on gamepad, and key inputs on keyboard, don't work.  I believe that somehow (in switching the Viewport controller class perhaps?), the default keys set up in Project Settings -> Inputs have ceased to apply, but I've yet to figure out how to re-apply them.  That's the next step.